sort wawrnings

+++ src/llvmcodegen.cc
@@ -135,7 +135,7 @@ Value *NodeDecl::llvm_codegen(LLVMCompiler *compiler)
 
     AllocaInst *alloc = temp_builder.CreateAlloca(compiler->builder.getInt32Ty(), 0, identifier);
 
-    if(scope<=compiler->locals[identifier].size())
+    if(scope<=(int)compiler->locals[identifier].size())
         compiler->locals[identifier][scope-1] = alloc;
     else
         compiler->locals[identifier].push_back(alloc);
